==> 04_Consolidado.ipynb <==
# NLP
import spacy

#other imports
import json
import pandas as pd

df2 = pd.read_excel("02_pagina12_parte1.xlsx")
df2['Diario'] = 'Pagina12'
df4 = pd.read_excel("02_pagina12_parte2.xlsx")
df4['Diario'] = 'Pagina12'
df3 = pd.read_excel("03_Infobae.xlsx")
df3['Diario'] = 'Infobae'

df = pd.concat([df2,df3,df4])
#df = df3
df['oraciones'] = None

nlp = spacy.load("es_core_news_lg") #es_core_news_lg modelo large, el transformer es es_dep_news_trf
def separate_sentences(text):
    doc = nlp(text)
    sentences = [sent.text for sent in doc.sents]
    return sentences

for index, row in df.iterrows():
    try:
        if row.oraciones is None:
            oracion = []
            oracion = separate_sentences(row.titulo)
            oracion.extend(separate_sentences(row.copete))
            oracion.extend(separate_sentences(row.contenido))
            oracion = json.dumps(oracion)
            df.at[index, 'oraciones'] = oracion
            print(f'Procesado Indice {index} de {len(df)}', end='\r')
    except:
        print(f'failed {index}')

df.drop_duplicates(subset=['link'], inplace=True)

df.to_excel('04_Consolidado.xlsx')
