==> 04_consolidar.ipynb <==
import pandas as pd

df2 = pd.read_excel("02_pagina12.xlsx")
df2['Diario'] = 'Pagina12'
df3 = pd.read_excel("03_Infobae.xlsx")
df3['Diario'] = 'Infobae'

df = pd.concat([df2,df3])

import nltk
from nltk.corpus import stopwords
from nltk.stem.wordnet import WordNetLemmatizer

# Gensim
import gensim
import gensim.corpora as corpora
from gensim.models import CoherenceModel

#other imports
import string
import re
import numpy as np
import pandas as pd

#nltk.download('stopwords')
#nltk.download('wordnet')
#nltk.download('vader_lexicon')

doc_complete = df['contenido']

stop = set(stopwords.words('spanish'))

exclude = set(string.punctuation)
lemma = WordNetLemmatizer()
def clean(doc):
    stop_free = " ".join([i for i in doc.lower().split() if i not in stop])
    punc_free = ''.join(ch for ch in stop_free if ch not in exclude)
    normalized = " ".join(lemma.lemmatize(word) for word in punc_free.split())
    return normalized

doc_clean = [clean(doc).split() for doc in doc_complete]  

dictionary = corpora.Dictionary(doc_clean)

doc_term_matrix = [dictionary.doc2bow(doc) for doc in doc_clean]

Lda = gensim.models.ldamodel.LdaModel
#Lda = gensim.models.ldamulticore.LdaModel


ldamodel = Lda(doc_term_matrix, num_topics=16, id2word = dictionary, passes=50)

df_lda = pd.DataFrame(ldamodel.print_topics(num_topics=16, num_words=5))

from nltk.sentiment import SentimentIntensityAnalyzer

sid = SentimentIntensityAnalyzer()

for i in range(0,1000):
    text = str(df.at[i, 'copete'])
    scores = sid.polarity_scores(text)
    print(scores)

import spacy
from transformers import pipeline

# Load the SpaCy English model
nlp = spacy.load("es_dep_news_trf") #es_core_news_lg modelo large, el transformer es es_dep_news_trf
# Define a text for sentiment analysis
text = "Esto es un terrible producto"
# Create a sentiment analysis pipeline using a pre-trained model
sentiment_analysis = pipeline("sentiment-analysis")
# Perform sentiment analysis on the text
result = sentiment_analysis(text)
# Print the sentiment result
print("Sentiment:", result[0]['label'])


for i in range(0,20):
    text = str(df.at[i, 'copete'])
    scores = sentiment_analysis(text)[0]
    print(f"Indice: {i}\nTexto: {text:10} \nSentimiento: {scores['label']} Puntaje: {scores['score']}\n")
