==> 07_Graficos.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

df = pd.read_excel('06_Analisis.xlsx')
df.drop(columns=['Unnamed: 0','indice','link','contenido','copete','oraciones','eval','estrellas_totales'], inplace=True)
df['label_final'] = df['label_final'].astype('category')
df['Diario'] = df['Diario'].astype('category')

df.head(10)

df.dtypes

# df['yearmonth'] = df['fecha'].dt.strftime('%m')

# # Create a count plot using Seaborn
# plt.figure(figsize=(10, 6))
# sns.countplot(x='yearmonth', hue='label_final', data=df)
# plt.title('Count of Rows per Category over Weeks')
# plt.xlabel('Week')
# plt.ylabel('Count')
# plt.legend(title='Category')
# plt.show()

sns.set_theme()
